Update reporters. Remove temporary JSON reporter for tests. Move screenshots dir to top level. Bump version.
