+++ test/functional/fixtures/screenshots-on-fails/test.js
@@ -93,7 +93,7 @@ if (config.useLocalBrowsers) {
                     expect(assertionHelper.isScreenshotDirExists()).eql(false);
                     assertionHelper.errorInEachBrowserNotContains(errs, SCREENSHOT_PATH_MESSAGE_TEXT, 0);
                     expect(testReport.warnings).eql([
-                        'Cannot take screenshots because the screenshot directory is not specified. To specify it, ' +
+                        'Was unable to take screenshots because the screenshot directory is not specified. To specify it, ' +
                         'use the "-s" or "--screenshots" command line option or the "screenshots" method of the ' +
                         'test runner in case you are using API.'
                     ]);
