+++ test/functional/setup.js
@@ -89,57 +89,6 @@ function closeLocalBrowsers () {
     return Promise.all(closeBrowserPromises);
 }
 
-// TODO remove once we've update reporters for #671
-var TempJSONReporter = function () {
-    return {
-        noColors:       true,
-        currentFixture: null,
-
-        report: {
-            startTime:  null,
-            endTime:    null,
-            userAgents: null,
-            passed:     0,
-            total:      0,
-            fixtures:   []
-        },
-
-        reportTaskStart: function reportTaskStart (startTime, userAgents, testCount) {
-            this.report.startTime  = startTime;
-            this.report.userAgents = userAgents;
-            this.report.total      = testCount;
-        },
-
-        reportFixtureStart: function reportFixtureStart (name, fixturePath) {
-            this.currentFixture = { name: name, path: fixturePath, tests: [] };
-            this.report.fixtures.push(this.currentFixture);
-        },
-
-        reportTestDone: function reportTestDone (name, testRunInfo) {
-            var _this = this;
-
-            var errs = testRunInfo.errs.map(function (err) {
-                return _this.formatError(err);
-            });
-
-            this.currentFixture.tests.push({
-                name:           name,
-                errs:           errs,
-                durationMs:     testRunInfo.durationMs,
-                unstable:       testRunInfo.unstable,
-                screenshotPath: testRunInfo.screenshotPath
-            });
-        },
-
-        reportTaskDone: function reportTaskDone (endTime, passed, warnings) {
-            this.report.passed   = passed;
-            this.report.endTime  = endTime;
-            this.report.warnings = warnings;
-
-            this.write(JSON.stringify(this.report, null, 2));
-        }
-    };
-};
 
 before(function () {
     var mocha = this;
@@ -198,7 +147,7 @@ before(function () {
                     .filter(function (test) {
                         return testName ? test === testName : true;
                     })
-                    .reporter(TempJSONReporter, {
+                    .reporter('json', {
                         write: function (data) {
                             report += data;
                         },
